Moved feed checking to separate worker processes/threads. This means individual feeds are no longer blocking.
This helps by:
a) reducing individual TTFB and response waits by up to 4 (the number of workers)
b) not blocking other feeds when one feed is taking a long time

For example, if we have 20 feeds and 4 workers, then the first 4 feeds
will all be queried and retrieved simultaneously. So the wait times
overlap rather than occurring one after the other, as do the responses.

Further to this example, let's say we have 2 feeds that take a long time
to respond. In this case, two workers will continue on regardless, and two
will wait on those slow feeds. So the slow feeds overlap, and they don't
block other feeds from running.