descriptions, cache, config, error Added feed descriptions, courtesy of: https://github.com/bibendi/Feedindicator/commit/e16a23e57bc6d38ae27dff655aefe7c11d374d6c
Added a config option to toggle on/off the above function.

Added the creation of the cache dir to the install script.
This prevents the below error from occurring.

Added error handling for RSS feed icon downloads.
If the cache dir didn't exist or was not writeable, the icons would
fail to download, thus causing the entire feed to abort rather than
just downloading the feeds without their icons.